+++ concrete_datastore/api/v1/views.py
@@ -55,6 +55,10 @@ from concrete_datastore.concrete.models import (  # pylint:disable=E0611
     Email,
     SecureConnectToken,
 )
+from concrete_datastore.api.v1.throttling import (
+    CustomUserRateThrottle,
+    CustomAnonymousRateThrottle,
+)
 from concrete_datastore.api.v1.permissions import (
     UserAccessPermission,
     filter_queryset_by_permissions,
@@ -1526,6 +1530,11 @@ class ApiModelViewSet(PaginatedViewSet, viewsets.ModelViewSet):
         URLTokenExpiryAuthentication,
     )
 
+    def get_throttles(self):
+        if settings.ENABLE_THROTTLING is False:
+            return []
+        return (CustomUserRateThrottle(), CustomAnonymousRateThrottle())
+
     def dispatch(self, request, *args, **kwargs):
         # Retrieve response first then log
         rsp = super(ApiModelViewSet, self).dispatch(request, *args, **kwargs)
@@ -1808,8 +1817,9 @@ class ApiModelViewSet(PaginatedViewSet, viewsets.ModelViewSet):
         return self.serializer_class_nested
 
     def perform_create(self, serializer):
-
-        attrs = {'created_by': self.request.user}
+        attrs = {}
+        if self.request.user.is_authenticated:
+            attrs.update({'created_by': self.request.user})
 
         try:
             divider = self.get_divider()